backend/requirements.txt
Django
djangorestframework
gunicorn
mysqlclient
python-dotenv
GitPython==3.1.40
ruff
isort
django-cors-headers==4.7.0
requests
cryptography
dj-rest-auth
django-allauth
PyJWT
djangorestframework-simplejwt
rest_framework_simplejwt
dj-rest-auth[with_social]
django-extensions
Pillow
